requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pillow==10.3.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
pytesseract